rename salah file
